+++ shared/src/test/scala/org/homermultitext/edmodel/MidOrthographySpec.scala
@@ -5,11 +5,10 @@ class MidOrthographySpec extends FlatSpec {
 
    "The library" should "include case objects for all edition types" in {
      assert(HmtNamedEntityEdition.label == "named entities")
-     assert(HmtNamedEntityEdition.label == "named entities")
      assert (HmtDiplomaticEdition.label ==  "diplomatic edition")
      assert (HmtScribalNormalizedEdition.label ==  "scribal edition")
      assert (HmtEditorsNormalizedEdition.label ==  "editorially normalized edition")
    }
 
-   
+
 }